Ignore unwatched directories when detecting a change
This is a problem that happens only on windows: because some change may be
missing information we have to go through `detectFileChange` algorithm
that relies on mtime to decide what changed. Now, if you ignore `dot`
files when watching and then an event is emitted on a "dot" directory
that changed we'll go through `detectFileChange` and through when we
don't find the dir in `dirRegistery`.

We solve this by just returning, hence ignoring the change event.