Workaround race condition
Between the time we get to call lstat on a file that was changed,
the file maybe deleted and lstat will return an error. This is
currently causing programs to crash (if they don't listent to the
error event, most don't). I'm seeing this issue with fly check
on emacs. If you change a file quickly it will update the fly_check
file by deleting and writing again.

The sane API supplies clients with node stat objects, so emitting
the event without the stat object is out of the question.
The easiest thing to do is to ignore the event, now the problem
becomes is inconsistency because we'll get a delete event from watchman
(either before or after). If clients maintain a state of the world,
that means they can get confused by getting a delete event for a
file that was never added.

However, I can't think of any other reliable solution and hoping
clients would be fine with that.